requirements.txt
Django==5.0.4
channels==4.1.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.34
Pillow==10.3.0